collision

+++ src/xpbd/src/lib.rs
@@ -5,4 +5,4 @@ mod time_manager;
 pub mod world;
 
 pub type V2 = nalgebra::Vector2<f32>;
-pub type C2 = nalgebra::Vector2<isize>;
+pub type C2 = nalgebra::Vector2<i32>;
